FILE: moves app-migration
